Files are separated into different packages
