+++ app/src/main/java/com/example/bookofrecipes/fragments/SettingsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.bookofrecipes
+package com.example.bookofrecipes.fragments
 
 import android.os.Bundle
 import android.view.View
