+++ app/src/main/res/layout/fragment_home_advanced.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context = ".HomeAdvancedFragment">
+    tools:context = ".fragments.HomeAdvancedFragment">
 
     <ImageView
         android:id="@+id/ivAboutImage"
